test: Test that direction changes when onkeypress
This checks if on keypress the snake drection changes to the
corresponding arrow key. i.e if right arrow key is pressed the snake
direction is right. It prevents the snake direction from changing
if the snake is already going in the direction of the key pressed.
It prevents the snake direction from changing if the snake's head hasn't
moved up to the snake's thickness. This prevents weird rendering issues
where the snake moves into itself.