feat: Snake direction can be changed
- The Snake.changeDirection method allows for the snake's direction to
be changed
- The Snake.changeDirection method is called on the snake component
whenever the "W A S D" keys are pressed. Snake.changeDirection is called
with "up" when the "W" key is pressed, with "down" when the "S" key is
pressed, with "right" when the "D" key is pressed, with "left" when the
"A" key is pressed,
- A getter method was added to the Snake class to for getting a copy of
the snake body
- The move method was updated to remove the tail when its length is
equal the snake's thickness.