feat: Run render functions when added
This iterates on when single render functions are run. Before they were
run when the canvas was mounted or resized. With this update they are
run the same as before and now they are run once they are added to the
queue if the canvas has been mounted, if not they are added to the queue
and run when the canvas is mounted.