test: Migrate tests to use vitest
This updates the tests to use vitest globals instead of jest. It removes
tests on the Snake class (I felt like this was an implementation detail
and as such it didn't need to be tested, this also allows me to implent
the snake movement in other ways that don't require the class). It
replaces manual repetitive tests on snake movement in different
directions with test.each