refactor: Change how the snake works under the hood
Previously all the snake data (it's position, the head, the tail etc.)
and logic (it's motion logic (moving the snake, changing it's direction)
and it's rendering logic (drawing the snake, clearing the drawing))
was stored in a class.
Then on each animation frame the snake woud be rendered and the
application store updated with/based on the data in the class. However,
the snake class was becoming unwieldy, and it was difficulty to add new
features and debug existing features. So the class was scrapped. The
motion and rendering logic were moved into dedicated files, and the
applications store was used the single source of truth for the snake
data (it honestly made no sense to have it in multiple places).

The gameover logic (if the snake has bitten itself, gameover) was moved
from the Level component to the snake's animation function.