refactor: Change how the food works
Like the snake, the food data and some of it's rendering logic was
stored in a class. On it's own the class was not a problem. However,
there was also a store with the food position which the rest of the
application interacted with and whenever the food was re-rendered (when
the snake eats it), the store would have to be updated with the new
position. Just like with the snake it made no sense to have 2 copies of
the same data. So all that logic was moved to the Food component.

The logic for re-rendering the food (when the snake eats it) was shared
between the Level and Food components, this moves all the checks for
when the snake eats the food and the subsequent updates to the UI if it
has to the Snake component, and the because the refactor simpplified the
rendering and re-rendering logic in Food component, nothing had to be
done on that end.

Before this refactor, sometime the snake would eat only part of the food
or completely eat the food at the harder difficulties ("Nightmare",
"Hell", "Beg For Mercy", "Cry To Your Mommy") and it wouldn't register
as being eaten (it wouldn't re-render), this fixes that.