fix: Fix animation bug on window resize
Previously, animations would play faster than normal whenever the window
was resized or paused animations would play, then pause.

The problem?
When the widown resize event is fired, the canvas is resized &
rescaled, then animations are rerun. Why are animations rerun? Whenever
the canvas is scaled, it clears the entire canvas, hence, everything has
to be redrawn. The problem with this is, it happens everytime the window
is resized. Beacuse, resizing the window doesn't fire the event just
once isn't, the scaling-resizing-animating cycle happens many times

The solution?
Debounce the window resize event and only rescale & resize the
canvas once the use window resize event ends.